Check year range in both side
We cannot expect all users will always tag the files in the year tag with valid years. A lot of them will abuse this tag field to tag a full date in pure number, like 19991231. Such date is not parseable by DateTime.TryParse and is not within valid year range for DateTime constructor. We check if that is the case before creating a new DateTime instance.